agfawg

+++ Kirjastojärjestelmä/poisto.py
@@ -4,7 +4,7 @@ from tabulate import tabulate
 conn = sqlite3.connect('./Kannat/Kirjasto.db')
 cur = conn.cursor()
 def poisto():
-    poistovalinta = input("Haluatko poistaa (k)irjan vai (a)siakkaan: ").lower()
+    poistovalinta = input("Haluatko poistaa (k)irjan, (a)siakkaan vai (ad)minin: ").lower()
 
     if poistovalinta == "a":
                 print("")
@@ -15,12 +15,20 @@ def poisto():
                 print("")
                 poistoid = input('Anna poistettavan asiakkaan sähköpostiosoite: ')
                 print("")
-                cur.execute('DELETE FROM asiakkaat WHERE sähköpostiosoite = ?', (poistoid,))
-                conn.commit()
-                cur.execute('SELECT * FROM asiakkaat')
-                data = cur.fetchall()
-                table = tabulate(data)
-                print(table)
+                cur.execute("SELECT sakkosaldo FROM asiakkaat WHERE sähköpostiosoite = ?", (poistoid,))
+                sak = cur.fetchone()
+
+                if sak[0] > 0:
+                    print("et voi poistaa käyttäjää joilla on sakkoja maksamatta\n")
+                elif sak is None:
+                        print("asiakasta ei löytynyt")
+                else:
+                    cur.execute('DELETE FROM asiakkaat WHERE sähköpostiosoite = ?', (poistoid,))
+                    conn.commit()
+                    cur.execute('SELECT * FROM asiakkaat')
+                    data = cur.fetchall()
+                    table = tabulate(data)
+                    print(table)
 
     elif poistovalinta == "k":
                 print("")
@@ -38,5 +46,23 @@ def poisto():
                 table = tabulate(data)
                 print(table)
 
+    elif poistovalinta == "ad":
+        print("")
+        cur.execute('SELECT * FROM admin')
+        data = cur.fetchall()
+        table = tabulate(data)
+        print(table)
+        print("")
+        poistoid = input('Anna poistettavan adminin sähköpostiosoite: ')
+        print("")
+        cur.execute('DELETE FROM admin WHERE sähköpostiosoite = ?', (poistoid,))
+        conn.commit()
+        cur.execute('SELECT * FROM admin')
+        data = cur.fetchall()
+        table = tabulate(data)
+        print(table)
+
     else:
                 print("Kirjoita joko a tai k! ")
+
+
